requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0
python-multipart>=0.0.9
PyJWT[crypto]>=2.8.0
